requirements.txt
langchain
langchain-openai
gradio
chromadb
tiktoken
pypdf
python-dotenv
scikit-learn
pandas
umap-learn
numpy
flashrank
unstructured[all-docs]
pillow
langchain-experimental